Before parser WallWextraWerror sharpening

+++ parser/preparser.c
@@ -31,7 +31,7 @@ static int	is_empty(const char *str)
 	while (str[++i])
 		if (!ft_isspace(str[i]))
 			break ;
-	if (i == (int)ft_strlen(str))
+	if (i == ft_strlen(str))
 		return (1);
 	return (0);
 }
